refactor: 기존 어드민 패키지를 auth 패키지에 포함되도록 변경
issue#6
