+++ wabi/src/main/kotlin/com/wap/wabi/auth/admin/entity/Enum/AdminStatus.java
@@ -1,4 +1,4 @@
-package com.wap.wabi.admin.entity.Enum;
+package com.wap.wabi.auth.admin.entity.Enum;
 
 public enum AdminStatus {
     ACTIVE,
