+++ wabi/src/main/kotlin/com/wap/wabi/auth/admin/payload/request/AdminLoginRequest.kt
@@ -1,0 +1,6 @@
+package com.wap.wabi.auth.admin.payload.request
+
+data class AdminLoginRequest(
+    val id: String,
+    val password: String
+)
