+++ wabi/src/main/kotlin/com/wap/wabi/auth/admin/service/JwtService.kt
@@ -1,0 +1,8 @@
+package com.wap.wabi.auth.admin.service
+
+import org.springframework.stereotype.Component
+
+@Component
+class JwtService {
+
+}
